refactor: 에러코드 반영 (#20)
