+++ src/main/java/com/example/i_commerce/domain/order/service/DeliveryService.java
@@ -5,12 +5,12 @@ import com.example.i_commerce.domain.order.entity.Order;
 import com.example.i_commerce.domain.order.entity.OrderProduct;
 import com.example.i_commerce.domain.order.entity.emuns.DeliveryStatus;
 import com.example.i_commerce.domain.order.event.dto.PaymentCompletedEvent;
+import com.example.i_commerce.domain.order.exception.OrderErrorCode;
 import com.example.i_commerce.domain.order.repository.DeliveryRepository;
 import com.example.i_commerce.domain.order.repository.OrderRepository;
 import com.example.i_commerce.domain.product.entity.ProductItem;
 import com.example.i_commerce.domain.product.repository.ProductItemRepository;
-import com.example.i_commerce.global.error.AppException;
-import com.example.i_commerce.global.error.ErrorCode;
+import com.example.i_commerce.global.exception.AppException;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -28,7 +28,7 @@ public class DeliveryService {
 
     public void createDelivery(PaymentCompletedEvent event) {
 
-        Order order = orderRepository.findById(event.orderId()).orElseThrow(() -> new AppException(ErrorCode.ORDER_TEMP_ERROR));
+        Order order = orderRepository.findById(event.orderId()).orElseThrow(() -> new AppException(OrderErrorCode.ORDER_TEMP_ERROR));
 
         List<Long> productIds = order.getOrderProducts().stream()
                 .map(OrderProduct::getProductSkuId).toList();
