+++ src/main/java/com/example/i_commerce/domain/order/service/PaymentService.java
@@ -4,9 +4,9 @@ import com.example.i_commerce.domain.order.entity.Payment;
 import com.example.i_commerce.domain.order.entity.emuns.PaymentStatus;
 import com.example.i_commerce.domain.order.event.dto.OrderCreatedEvent;
 import com.example.i_commerce.domain.order.event.dto.PaymentCompletedEvent;
+import com.example.i_commerce.domain.order.exception.OrderErrorCode;
 import com.example.i_commerce.domain.order.repository.PaymentRepository;
-import com.example.i_commerce.global.error.AppException;
-import com.example.i_commerce.global.error.ErrorCode;
+import com.example.i_commerce.global.exception.AppException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -20,7 +20,7 @@ public class PaymentService {
 
     public void pay(OrderCreatedEvent event) {
         //TODO : PG연결
-        Payment payment = paymentRepository.findById(event.paymentId()).orElseThrow(() -> new AppException(ErrorCode.ORDER_TEMP_ERROR));
+        Payment payment = paymentRepository.findById(event.paymentId()).orElseThrow(() -> new AppException(OrderErrorCode.ORDER_TEMP_ERROR));
         payment.changePayStatus(PaymentStatus.PAID);
 
         publisher.publishEvent(new PaymentCompletedEvent(event.orderId(), event.paymentId(), event.userId()));
